Juuri luotu startScreen poistettu ja alkunäkymä luotu suoraan App.js tiedostoon
